simulator-python/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
faker>=22.0.0
